+++ embassy-executor/tests/test.rs
@@ -40,9 +40,6 @@ fn setup() -> (&'static Executor, Trace) {
     let trace = Trace::new();
     let context = Box::leak(Box::new(trace.clone())) as *mut _ as *mut ();
     let executor = &*Box::leak(Box::new(Executor::new(context)));
-    unsafe {
-        executor.initialize();
-    }
 
     (executor, trace)
 }
@@ -153,3 +150,7 @@ fn executor_task_cfg_args() {
         let (_, _, _) = (a, b, c);
     }
 }
+
+// We need this for the test to compile, even though we don't want to use timers at the moment.
+#[no_mangle]
+fn _embassy_time_schedule_wake(_at: u64, _waker: &core::task::Waker) {}