+++ embassy-executor/CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-- `raw::Executor` now has an `fn initialize` that must be called once before starting to poll it.
+- embassy-executor no longer provides an `embassy-time-queue-driver` implementation
+- Added `TaskRef::executor` to obtain a reference to a task's executor
+- integrated-timers are no longer processed when polling the executor.
 
 ## 0.6.3 - 2024-11-12