+++ embassy-time/CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- The `generic-queue` and related features have been removed (moved to embassy-time-queue-driver)
+- embassy-time no longer provides an `embassy-time-queue-driver` implementation
+
 ## 0.3.2 - 2024-08-05
 
 - Implement with_timeout()/with_deadline() method style call on Future
-- Add collapse_debuginfo to fmt.rs macros. 
+- Add collapse_debuginfo to fmt.rs macros.
 
 ## 0.3.1 - 2024-01-11